fix(groq): consolidate event handling and streamline attribute setting
- Removed the event_handler module and integrated its functionality into event_emitter.
- Introduced emit_message_events, emit_choice_events, and emit_streaming_response_events for better event management.
- Updated span_utils to utilize new attribute setting functions for model input and response.
- Simplified the configuration by removing the event_logger attribute from Config.
- Adjusted tests to reflect changes in event emission and attribute handling.
- Enhanced utility functions to improve clarity and maintainability.
