+++ packages/opentelemetry-instrumentation-groq/opentelemetry/instrumentation/groq/__init__.py
@@ -1,38 +1,35 @@
 """OpenTelemetry Groq instrumentation"""
 
-import json
 import logging
 import os
 import time
 from typing import Callable, Collection, Union
 
 from opentelemetry import context as context_api
-from opentelemetry._events import Event, EventLogger, get_event_logger
+from opentelemetry._events import EventLogger, get_event_logger
 from opentelemetry.instrumentation.groq.config import Config
-from opentelemetry.instrumentation.groq.event_handler import (
-    ChoiceEvent,
-    MessageEvent,
-    emit_event,
+from opentelemetry.instrumentation.groq.event_emitter import (
+    emit_choice_events,
+    emit_message_events,
+    emit_streaming_response_events,
+)
+from opentelemetry.instrumentation.groq.span_utils import (
+    set_input_attributes,
+    set_model_input_attributes,
+    set_model_response_attributes,
+    set_model_streaming_response_attributes,
+    set_response_attributes,
+    set_streaming_response_attributes,
 )
 from opentelemetry.instrumentation.groq.utils import (
-    dont_throw,
     error_metrics_attributes,
-    is_content_enabled,
-    model_as_dict,
-    set_span_attribute,
     shared_metrics_attributes,
-    should_send_prompts,
+    should_emit_events,
 )
 from opentelemetry.instrumentation.groq.version import __version__
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY, unwrap
 from opentelemetry.metrics import Counter, Histogram, Meter, get_meter
-from opentelemetry.semconv._incubating.attributes import (
-    gen_ai_attributes as GenAIAttributes,
-)
-from opentelemetry.semconv._incubating.attributes.gen_ai_attributes import (
-    GEN_AI_RESPONSE_ID,
-)
 from opentelemetry.semconv_ai import (
     SUPPRESS_LANGUAGE_MODEL_INSTRUMENTATION_KEY,
     LLMRequestTypeValues,
@@ -44,13 +41,11 @@ from opentelemetry.trace.status import Status, StatusCode
 from wrapt import wrap_function_wrapper
 
 from groq._streaming import AsyncStream, Stream
-from groq.types.chat.chat_completion import ChatCompletion
 
 logger = logging.getLogger(__name__)
 
 _instruments = ("groq >= 0.9.0",)
 
-CONTENT_FILTER_KEY = "content_filter_results"
 
 WRAPPED_METHODS = [
     {
@@ -74,192 +69,16 @@ def is_streaming_response(response):
     return isinstance(response, Stream) or isinstance(response, AsyncStream)
 
 
-def _dump_content(content):
-    if isinstance(content, str):
-        return content
-    json_serializable = []
-    for item in content:
-        if item.get("type") == "text":
-            json_serializable.append({"type": "text", "text": item.get("text")})
-        elif item.get("type") == "image":
-            json_serializable.append(
-                {
-                    "type": "image",
-                    "source": {
-                        "type": item.get("source").get("type"),
-                        "media_type": item.get("source").get("media_type"),
-                        "data": str(item.get("source").get("data")),
-                    },
-                }
-            )
-    return json.dumps(json_serializable)
-
-
-@dont_throw
-def _set_input_attributes(span, kwargs):
-    set_span_attribute(span, SpanAttributes.LLM_REQUEST_MODEL, kwargs.get("model"))
-    set_span_attribute(
-        span, SpanAttributes.LLM_REQUEST_MAX_TOKENS, kwargs.get("max_tokens_to_sample")
-    )
-    set_span_attribute(
-        span, SpanAttributes.LLM_REQUEST_TEMPERATURE, kwargs.get("temperature")
-    )
-    set_span_attribute(span, SpanAttributes.LLM_REQUEST_TOP_P, kwargs.get("top_p"))
-    set_span_attribute(
-        span, SpanAttributes.LLM_FREQUENCY_PENALTY, kwargs.get("frequency_penalty")
-    )
-    set_span_attribute(
-        span, SpanAttributes.LLM_PRESENCE_PENALTY, kwargs.get("presence_penalty")
-    )
-    set_span_attribute(
-        span, SpanAttributes.LLM_IS_STREAMING, kwargs.get("stream") or False
-    )
-
-    if should_send_prompts():
-        if kwargs.get("prompt") is not None:
-            set_span_attribute(
-                span, f"{SpanAttributes.LLM_PROMPTS}.0.user", kwargs.get("prompt")
-            )
-
-        elif kwargs.get("messages") is not None:
-            for i, message in enumerate(kwargs.get("messages")):
-                set_span_attribute(
-                    span,
-                    f"{SpanAttributes.LLM_PROMPTS}.{i}.content",
-                    _dump_content(message.get("content")),
-                )
-                set_span_attribute(
-                    span, f"{SpanAttributes.LLM_PROMPTS}.{i}.role", message.get("role")
-                )
-
-
-def _set_completions(span, choices):
-    if choices is None:
-        return
-
-    for choice in choices:
-        index = choice.get("index")
-        prefix = f"{SpanAttributes.LLM_COMPLETIONS}.{index}"
-        set_span_attribute(span, f"{prefix}.finish_reason", choice.get("finish_reason"))
-
-        if choice.get("content_filter_results"):
-            set_span_attribute(
-                span,
-                f"{prefix}.{CONTENT_FILTER_KEY}",
-                json.dumps(choice.get("content_filter_results")),
-            )
-
-        if choice.get("finish_reason") == "content_filter":
-            set_span_attribute(span, f"{prefix}.role", "assistant")
-            set_span_attribute(span, f"{prefix}.content", "FILTERED")
-
-            return
-
-        message = choice.get("message")
-        if not message:
-            return
-
-        set_span_attribute(span, f"{prefix}.role", message.get("role"))
-        set_span_attribute(span, f"{prefix}.content", message.get("content"))
-
-        function_call = message.get("function_call")
-        if function_call:
-            set_span_attribute(
-                span, f"{prefix}.tool_calls.0.name", function_call.get("name")
-            )
-            set_span_attribute(
-                span,
-                f"{prefix}.tool_calls.0.arguments",
-                function_call.get("arguments"),
-            )
-
-        tool_calls = message.get("tool_calls")
-        if tool_calls:
-            for i, tool_call in enumerate(tool_calls):
-                function = tool_call.get("function")
-                set_span_attribute(
-                    span,
-                    f"{prefix}.tool_calls.{i}.id",
-                    tool_call.get("id"),
-                )
-                set_span_attribute(
-                    span,
-                    f"{prefix}.tool_calls.{i}.name",
-                    function.get("name"),
-                )
-                set_span_attribute(
-                    span,
-                    f"{prefix}.tool_calls.{i}.arguments",
-                    function.get("arguments"),
-                )
-
-
-@dont_throw
-def _set_response_attributes(span, response, token_histogram):
-    response = model_as_dict(response)
-    set_span_attribute(span, SpanAttributes.LLM_RESPONSE_MODEL, response.get("model"))
-    set_span_attribute(span, GEN_AI_RESPONSE_ID, response.get("id"))
-
-    usage = response.get("usage") or {}
-    prompt_tokens = usage.get("prompt_tokens")
-    completion_tokens = usage.get("completion_tokens")
-    if usage:
-        set_span_attribute(
-            span, SpanAttributes.LLM_USAGE_TOTAL_TOKENS, usage.get("total_tokens")
-        )
-        set_span_attribute(
-            span, SpanAttributes.LLM_USAGE_COMPLETION_TOKENS, completion_tokens
-        )
-        set_span_attribute(span, SpanAttributes.LLM_USAGE_PROMPT_TOKENS, prompt_tokens)
-
-    if (
-        isinstance(prompt_tokens, int)
-        and prompt_tokens >= 0
-        and token_histogram is not None
-    ):
-        token_histogram.record(
-            prompt_tokens,
-            attributes={
-                SpanAttributes.LLM_TOKEN_TYPE: "input",
-                SpanAttributes.LLM_RESPONSE_MODEL: response.get("model"),
-            },
-        )
-
-    if (
-        isinstance(completion_tokens, int)
-        and completion_tokens >= 0
-        and token_histogram is not None
-    ):
-        token_histogram.record(
-            completion_tokens,
-            attributes={
-                SpanAttributes.LLM_TOKEN_TYPE: "output",
-                SpanAttributes.LLM_RESPONSE_MODEL: response.get("model"),
-            },
-        )
-
-    choices = response.get("choices")
-    if should_send_prompts() and choices:
-        _set_completions(span, choices)
-
-
-def _with_tracer_wrapper(func):
-    """Helper for providing tracer for wrapper functions."""
-
-    def _with_tracer(tracer, to_wrap):
-        def wrapper(wrapped, instance, args, kwargs):
-            return func(tracer, to_wrap, wrapped, instance, args, kwargs)
-
-        return wrapper
-
-    return _with_tracer
-
-
 def _with_chat_telemetry_wrapper(func):
     """Helper for providing tracer for wrapper functions. Includes metric collectors."""
 
     def _with_chat_telemetry(
-        tracer, token_histogram, choice_counter, duration_histogram, to_wrap
+        tracer,
+        token_histogram,
+        choice_counter,
+        duration_histogram,
+        event_logger,
+        to_wrap,
     ):
         def wrapper(wrapped, instance, args, kwargs):
             return func(
@@ -267,6 +86,7 @@ def _with_chat_telemetry_wrapper(func):
                 token_histogram,
                 choice_counter,
                 duration_histogram,
+                event_logger,
                 to_wrap,
                 wrapped,
                 instance,
@@ -318,45 +138,19 @@ def _process_streaming_chunk(chunk):
     return content, finish_reason, usage
 
 
-def _set_streaming_response_attributes(
-    span, accumulated_content, finish_reason=None, usage=None
+def _handle_streaming_response(
+    span, accumulated_content, finish_reason, usage, event_logger
 ):
-    """Set span attributes for accumulated streaming response."""
-    if not span.is_recording():
-        return
-
-    prefix = f"{SpanAttributes.LLM_COMPLETIONS}.0"
-    set_span_attribute(span, f"{prefix}.role", "assistant")
-    set_span_attribute(span, f"{prefix}.content", accumulated_content)
-    if finish_reason:
-        set_span_attribute(span, f"{prefix}.finish_reason", finish_reason)
-
-    if usage:
-        set_span_attribute(
-            span, SpanAttributes.LLM_USAGE_COMPLETION_TOKENS, usage.completion_tokens
-        )
-        set_span_attribute(
-            span, SpanAttributes.LLM_USAGE_PROMPT_TOKENS, usage.prompt_tokens
-        )
-        set_span_attribute(
-            span, SpanAttributes.LLM_USAGE_TOTAL_TOKENS, usage.total_tokens
+    set_model_streaming_response_attributes(span, usage)
+    if should_emit_events() and event_logger:
+        emit_streaming_response_events(accumulated_content, finish_reason, event_logger)
+    else:
+        set_streaming_response_attributes(
+            span, accumulated_content, finish_reason, usage
         )
 
 
-def _emit_streaming_response_events(
-    accumulated_content: str, finish_reason: Union[str, None]
-):
-    """Emit events for streaming response."""
-    emit_event(
-        ChoiceEvent(
-            index=0,
-            message={"content": accumulated_content, "role": "assistant"},
-            finish_reason=finish_reason or "unknown",
-        )
-    )
-
-
-def _create_stream_processor(response, span):
+def _create_stream_processor(response, span, event_logger):
     """Create a generator that processes a stream while collecting telemetry."""
     accumulated_content = ""
     finish_reason = None
@@ -372,17 +166,17 @@ def _create_stream_processor(response, span):
             usage = chunk_usage
         yield chunk
 
+    _handle_streaming_response(
+        span, accumulated_content, finish_reason, usage, event_logger
+    )
+
     if span.is_recording():
-        _set_streaming_response_attributes(
-            span, accumulated_content, finish_reason, usage
-        )
         span.set_status(Status(StatusCode.OK))
 
-    _emit_streaming_response_events(accumulated_content, finish_reason)
     span.end()
 
 
-async def _create_async_stream_processor(response, span):
+async def _create_async_stream_processor(response, span, event_logger):
     """Create an async generator that processes a stream while collecting telemetry."""
     accumulated_content = ""
     finish_reason = None
@@ -398,38 +192,30 @@ async def _create_async_stream_processor(response, span):
             usage = chunk_usage
         yield chunk
 
+    _handle_streaming_response(
+        span, accumulated_content, finish_reason, usage, event_logger
+    )
+
     if span.is_recording():
-        _set_streaming_response_attributes(
-            span, accumulated_content, finish_reason, usage
-        )
         span.set_status(Status(StatusCode.OK))
-    _emit_streaming_response_events(accumulated_content, finish_reason)
+
     span.end()
 
 
-@dont_throw
-def _emit_message_events(kwargs: dict):
-    for message in kwargs.get("messages", []):
-        emit_event(
-            MessageEvent(
-                content=message.get("content"), role=message.get("role", "unknown")
-            )
-        )
+def _handle_input(span, kwargs, event_logger):
+    set_model_input_attributes(span, kwargs)
+    if should_emit_events() and event_logger:
+        emit_message_events(kwargs, event_logger)
+    else:
+        set_input_attributes(span, kwargs)
 
 
-@dont_throw
-def _emit_choice_events(response: ChatCompletion):
-    for choice in response.choices:
-        emit_event(
-            ChoiceEvent(
-                index=choice.index,
-                message={
-                    "content": choice.message.content,
-                    "role": choice.message.role or "unknown",
-                },
-                finish_reason=choice.finish_reason,
-            )
-        )
+def _handle_response(span, response, token_histogram, event_logger):
+    set_model_response_attributes(span, response, token_histogram)
+    if should_emit_events() and event_logger:
+        emit_choice_events(response, event_logger)
+    else:
+        set_response_attributes(span, response)
 
 
 @_with_chat_telemetry_wrapper
@@ -438,6 +224,7 @@ def _wrap(
     token_histogram: Histogram,
     choice_counter: Counter,
     duration_histogram: Histogram,
+    event_logger: Union[EventLogger, None],
     to_wrap,
     wrapped,
     instance,
@@ -460,10 +247,7 @@ def _wrap(
         },
     )
 
-    if span.is_recording():
-        _set_input_attributes(span, kwargs)
-
-    _emit_message_events(kwargs)
+    _handle_input(span, kwargs, event_logger)
 
     start_time = time.time()
     try:
@@ -482,7 +266,7 @@ def _wrap(
 
     if is_streaming_response(response):
         try:
-            return _create_stream_processor(response, span)
+            return _create_stream_processor(response, span, event_logger)
         except Exception as ex:
             logger.warning(
                 "Failed to process streaming response for groq span, error: %s",
@@ -502,16 +286,13 @@ def _wrap(
                     attributes=metric_attributes,
                 )
 
-            if span.is_recording():
-                _set_response_attributes(span, response, token_histogram)
+            _handle_response(span, response, token_histogram, event_logger)
 
         except Exception as ex:  # pylint: disable=broad-except
             logger.warning(
                 "Failed to set response attributes for groq span, error: %s",
                 str(ex),
             )
-
-        _emit_choice_events(response)
 
         if span.is_recording():
             span.set_status(Status(StatusCode.OK))
@@ -525,6 +306,7 @@ async def _awrap(
     token_histogram: Histogram,
     choice_counter: Counter,
     duration_histogram: Histogram,
+    event_logger: Union[EventLogger, None],
     to_wrap,
     wrapped,
     instance,
@@ -546,16 +328,8 @@ async def _awrap(
             SpanAttributes.LLM_REQUEST_TYPE: LLMRequestTypeValues.COMPLETION.value,
         },
     )
-    try:
-        if span.is_recording():
-            _set_input_attributes(span, kwargs)
 
-    except Exception as ex:  # pylint: disable=broad-except
-        logger.warning(
-            "Failed to set input attributes for groq span, error: %s", str(ex)
-        )
-
-    _emit_message_events(kwargs)
+    _handle_input(span, kwargs, event_logger)
 
     start_time = time.time()
 
@@ -575,7 +349,7 @@ async def _awrap(
 
     if is_streaming_response(response):
         try:
-            return await _create_async_stream_processor(response, span)
+            return await _create_async_stream_processor(response, span, event_logger)
         except Exception as ex:
             logger.warning(
                 "Failed to process streaming response for groq span, error: %s",
@@ -594,14 +368,10 @@ async def _awrap(
                 attributes=metric_attributes,
             )
 
-        if span.is_recording():
-            _set_response_attributes(span, response, token_histogram)
+        _handle_response(span, response, token_histogram, event_logger)
 
         if span.is_recording():
             span.set_status(Status(StatusCode.OK))
-
-        _emit_choice_events(response)
-
     span.end()
     return response
 
@@ -650,9 +420,10 @@ class GroqInstrumentor(BaseInstrumentor):
                 duration_histogram,
             ) = (None, None, None)
 
+        event_logger = None
         if not Config.use_legacy_attributes:
             event_logger_provider = kwargs.get("event_logger_provider")
-            Config.event_logger = get_event_logger(
+            event_logger = get_event_logger(
                 __name__, __version__, event_logger_provider=event_logger_provider
             )
 
@@ -670,6 +441,7 @@ class GroqInstrumentor(BaseInstrumentor):
                         token_histogram,
                         choice_counter,
                         duration_histogram,
+                        event_logger,
                         wrapped_method,
                     ),
                 )
@@ -689,6 +461,7 @@ class GroqInstrumentor(BaseInstrumentor):
                         token_histogram,
                         choice_counter,
                         duration_histogram,
+                        event_logger,
                         wrapped_method,
                     ),
                 )
