+++ packages/opentelemetry-instrumentation-groq/opentelemetry/instrumentation/groq/config.py
@@ -1,6 +1,4 @@
-from typing import Callable, Optional
-
-from opentelemetry._events import EventLogger
+from typing import Callable
 
 
 class Config:
@@ -8,4 +6,3 @@ class Config:
     exception_logger = None
     get_common_metrics_attributes: Callable[[], dict] = lambda: {}
     use_legacy_attributes = True
-    event_logger: Optional[EventLogger] = None
