+++ packages/opentelemetry-instrumentation-groq/opentelemetry/instrumentation/groq/event_emitter.py
@@ -1,17 +1,19 @@
-from dataclasses import asdict, dataclass
+from dataclasses import asdict
 from enum import Enum
-from typing import Any, List, Literal, Optional, TypedDict, Union
+from typing import Union
 
-from opentelemetry._events import Event
+from opentelemetry._events import Event, EventLogger
+from opentelemetry.instrumentation.groq.event_models import ChoiceEvent, MessageEvent
 from opentelemetry.instrumentation.groq.utils import (
-    is_content_enabled,
+    dont_throw,
     should_emit_events,
+    should_send_prompts,
 )
 from opentelemetry.semconv._incubating.attributes import (
     gen_ai_attributes as GenAIAttributes,
 )
 
-from .config import Config
+from groq.types.chat.chat_completion import ChatCompletion
 
 
 class Roles(Enum):
@@ -30,64 +32,69 @@ EVENT_ATTRIBUTES = {
 """The attributes to be used for the event."""
 
 
-class _FunctionToolCall(TypedDict):
-    function_name: str
-    arguments: Optional[dict[str, Any]]
+@dont_throw
+def emit_message_events(kwargs: dict, event_logger):
+    for message in kwargs.get("messages", []):
+        emit_event(
+            MessageEvent(
+                content=message.get("content"), role=message.get("role", "unknown")
+            ),
+            event_logger=event_logger,
+        )
 
 
-class ToolCall(TypedDict):
-    """Represents a tool call in the AI model."""
-
-    id: str
-    function: _FunctionToolCall
-    type: Literal["function"]
-
-
-class CompletionMessage(TypedDict):
-    """Represents a message in the AI model."""
-
-    content: Any
-    role: str = "assistant"
-
-
-@dataclass
-class MessageEvent:
-    """Represents an input event for the AI model."""
-
-    content: Any
-    role: str = "user"
-    tool_calls: Optional[List[ToolCall]] = None
+@dont_throw
+def emit_choice_events(response: ChatCompletion, event_logger):
+    for choice in response.choices:
+        emit_event(
+            ChoiceEvent(
+                index=choice.index,
+                message={
+                    "content": choice.message.content,
+                    "role": choice.message.role or "unknown",
+                },
+                finish_reason=choice.finish_reason,
+            ),
+            event_logger=event_logger,
+        )
 
 
-@dataclass
-class ChoiceEvent:
-    """Represents a completion event for the AI model."""
+@dont_throw
+def emit_streaming_response_events(
+    accumulated_content: str, finish_reason: Union[str, None], event_logger
+):
+    """Emit events for streaming response."""
+    emit_event(
+        ChoiceEvent(
+            index=0,
+            message={"content": accumulated_content, "role": "assistant"},
+            finish_reason=finish_reason or "unknown",
+        ),
+        event_logger,
+    )
 
-    index: int
-    message: CompletionMessage
-    finish_reason: str = "unknown"
-    tool_calls: Optional[List[ToolCall]] = None
 
-
-def emit_event(event: Union[MessageEvent, ChoiceEvent]) -> None:
+def emit_event(
+    event: Union[MessageEvent, ChoiceEvent], event_logger: Union[EventLogger, None]
+) -> None:
     """
     Emit an event to the OpenTelemetry SDK.
 
     Args:
         event: The event to emit.
     """
-    if not should_emit_events():
+    if not should_emit_events() or event_logger is None:
         return
 
     if isinstance(event, MessageEvent):
-        _emit_message_event(event)
+        _emit_message_event(event, event_logger)
     elif isinstance(event, ChoiceEvent):
-        _emit_choice_event(event)
+        _emit_choice_event(event, event_logger)
     else:
         raise TypeError("Unsupported event type")
 
 
-def _emit_message_event(event: MessageEvent) -> None:
+def _emit_message_event(event: MessageEvent, event_logger: EventLogger) -> None:
     body = asdict(event)
 
     if event.role in VALID_MESSAGE_ROLES:
@@ -105,16 +112,16 @@ def _emit_message_event(event: MessageEvent) -> None:
     elif event.tool_calls is None:
         del body["tool_calls"]
 
-    if not is_content_enabled():
+    if not should_send_prompts():
         del body["content"]
         if body.get("tool_calls") is not None:
             for tool_call in body["tool_calls"]:
                 tool_call["function"].pop("arguments", None)
 
-    Config.event_logger.emit(Event(name=name, body=body, attributes=EVENT_ATTRIBUTES))
+    event_logger.emit(Event(name=name, body=body, attributes=EVENT_ATTRIBUTES))
 
 
-def _emit_choice_event(event: ChoiceEvent) -> None:
+def _emit_choice_event(event: ChoiceEvent, event_logger: EventLogger) -> None:
     body = asdict(event)
     if event.message["role"] == Roles.ASSISTANT.value:
         # According to the semantic conventions, the role is conditionally required if available
@@ -124,12 +131,12 @@ def _emit_choice_event(event: ChoiceEvent) -> None:
     if event.tool_calls is None:
         del body["tool_calls"]
 
-    if not is_content_enabled():
+    if not should_send_prompts():
         body["message"].pop("content", None)
         if body.get("tool_calls") is not None:
             for tool_call in body["tool_calls"]:
                 tool_call["function"].pop("arguments", None)
 
-    Config.event_logger.emit(
+    event_logger.emit(
         Event(name="gen_ai.choice", body=body, attributes=EVENT_ATTRIBUTES)
     )
