+++ packages/opentelemetry-instrumentation-groq/opentelemetry/instrumentation/groq/utils.py
@@ -4,7 +4,6 @@ import traceback
 from importlib.metadata import version
 
 from opentelemetry import context as context_api
-from opentelemetry._events import EventLogger
 from opentelemetry.instrumentation.groq.config import Config
 from opentelemetry.semconv_ai import SpanAttributes
 
@@ -13,9 +12,7 @@ GEN_AI_SYSTEM_GROQ = "groq"
 
 _PYDANTIC_VERSION = version("pydantic")
 
-OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT = (
-    "OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT"
-)
+TRACELOOP_TRACE_CONTENT = "TRACELOOP_TRACE_CONTENT"
 
 
 def set_span_attribute(span, name, value):
@@ -25,7 +22,7 @@ def set_span_attribute(span, name, value):
 
 def should_send_prompts():
     return (
-        os.getenv("TRACELOOP_TRACE_CONTENT") or "true"
+        os.getenv(TRACELOOP_TRACE_CONTENT) or "true"
     ).lower() == "true" or context_api.get_value("override_enable_content_tracing")
 
 
@@ -86,20 +83,10 @@ def model_as_dict(model):
         return model
 
 
-def is_content_enabled() -> bool:
-    capture_content = os.environ.get(
-        OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT, "false"
-    )
-
-    return capture_content.lower() == "true"
-
-
 def should_emit_events() -> bool:
     """
     Checks if the instrumentation isn't using the legacy attributes
     and if the event logger is not None.
     """
 
-    return not Config.use_legacy_attributes and isinstance(
-        Config.event_logger, EventLogger
-    )
+    return not Config.use_legacy_attributes
