+++ packages/opentelemetry-instrumentation-groq/tests/traces/conftest.py
@@ -5,9 +5,7 @@ import os
 import pytest
 from groq import AsyncGroq, Groq
 from opentelemetry.instrumentation.groq import GroqInstrumentor
-from opentelemetry.instrumentation.groq.utils import (
-    OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT,
-)
+from opentelemetry.instrumentation.groq.utils import TRACELOOP_TRACE_CONTENT
 from opentelemetry.sdk._events import EventLoggerProvider
 from opentelemetry.sdk._logs import LoggerProvider
 from opentelemetry.sdk._logs.export import (
@@ -100,7 +98,7 @@ def instrument_legacy(reader, tracer_provider, meter_provider):
 def instrument_with_content(
     reader, tracer_provider, event_logger_provider, meter_provider
 ):
-    os.environ.update({OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT: "True"})
+    os.environ.update({TRACELOOP_TRACE_CONTENT: "True"})
 
     instrumentor = GroqInstrumentor(
         use_legacy_attributes=False,
@@ -114,7 +112,7 @@ def instrument_with_content(
 
     yield instrumentor
 
-    os.environ.pop(OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT, None)
+    os.environ.pop(TRACELOOP_TRACE_CONTENT, None)
     instrumentor.uninstrument()
 
 
@@ -122,7 +120,7 @@ def instrument_with_content(
 def instrument_with_no_content(
     reader, tracer_provider, event_logger_provider, meter_provider
 ):
-    os.environ.update({OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT: "False"})
+    os.environ.update({TRACELOOP_TRACE_CONTENT: "False"})
 
     instrumentor = GroqInstrumentor(
         use_legacy_attributes=False, enrich_token_usage=True
@@ -135,7 +133,7 @@ def instrument_with_no_content(
 
     yield instrumentor
 
-    os.environ.pop(OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT, None)
+    os.environ.pop(TRACELOOP_TRACE_CONTENT, None)
     instrumentor.uninstrument()
